LLAMA VISION/requirements.txt
openai
streamlit>=1.27.0
Pillow>=10.0.0
requests>=2.31.0
